% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{delete_batch_prediction}
\alias{delete_batch_prediction}
\title{Assigns the DELETED status to a BatchPrediction, rendering it unusable}
\usage{
delete_batch_prediction(BatchPredictionId)
}
\arguments{
\item{BatchPredictionId}{[required] A user-supplied ID that uniquely identifies the \code{BatchPrediction}.}
}
\description{
Assigns the DELETED status to a \code{BatchPrediction}, rendering it unusable.
}
\details{
After using the \code{DeleteBatchPrediction} operation, you can use the GetBatchPrediction operation to verify that the status of the \code{BatchPrediction} changed to DELETED.

\strong{Caution:} The result of the \code{DeleteBatchPrediction} operation is irreversible.
}
\section{Accepted Parameters}{
\preformatted{delete_batch_prediction(
  BatchPredictionId = "string"
)
}
}

