% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{delete_data_source}
\alias{delete_data_source}
\title{Assigns the DELETED status to a \code{DataSource}, rendering it unusable}
\usage{
delete_data_source(
  DataSourceId = "string"
)
}
\arguments{
\item{DataSourceId}{[required] A user-supplied ID that uniquely identifies the \code{DataSource}.}
}
\description{
Assigns the DELETED status to a \code{DataSource}, rendering it unusable.
}
\details{
After using the \code{DeleteDataSource} operation, you can use the GetDataSource operation to verify that the status of the \code{DataSource} changed to DELETED.

\strong{Caution:} The results of the \code{DeleteDataSource} operation are irreversible.
}
\examples{

}
