% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{delete_ml_model}
\alias{delete_ml_model}
\title{Assigns the DELETED status to an MLModel, rendering it unusable}
\usage{
delete_ml_model(MLModelId)
}
\arguments{
\item{MLModelId}{[required] A user-supplied ID that uniquely identifies the \code{MLModel}.}
}
\description{
Assigns the \code{DELETED} status to an \code{MLModel}, rendering it unusable.
}
\details{
After using the \code{DeleteMLModel} operation, you can use the \code{GetMLModel} operation to verify that the status of the \code{MLModel} changed to DELETED.

\strong{Caution:} The result of the \code{DeleteMLModel} operation is irreversible.
}
\section{Accepted Parameters}{
\preformatted{delete_ml_model(
  MLModelId = "string"
)
}
}

