% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Deletes the specified tags associated with an ML object}
\usage{
delete_tags(TagKeys, ResourceId, ResourceType)
}
\arguments{
\item{TagKeys}{One or more tags to delete.}

\item{ResourceId}{The ID of the tagged ML object. For example, \code{exampleModelId}.}

\item{ResourceType}{The type of the tagged ML object.}
}
\description{
Deletes the specified tags associated with an ML object. After this operation is complete, you can't recover deleted tags.
}
\details{
If you specify a tag that doesn't exist, Amazon ML ignores it.
}
\examples{

}
