% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{describe_evaluations}
\alias{describe_evaluations}
\title{Returns a list of DescribeEvaluations that match the search criteria in the request}
\usage{
describe_evaluations(FilterVariable = NULL, EQ = NULL, GT = NULL,
  LT = NULL, GE = NULL, LE = NULL, NE = NULL, Prefix = NULL,
  SortOrder = NULL, NextToken = NULL, Limit = NULL)
}
\arguments{
\item{FilterVariable}{Use one of the following variable to filter a list of \code{Evaluation} objects:
\itemize{
\item \code{CreatedAt} - Sets the search criteria to the \code{Evaluation} creation date.
\item \code{Status} - Sets the search criteria to the \code{Evaluation} status.
\item \code{Name} - Sets the search criteria to the contents of \code{Evaluation} \code{Name}.
\item \code{IAMUser} - Sets the search criteria to the user account that invoked an \code{Evaluation}.
\item \code{MLModelId} - Sets the search criteria to the \code{MLModel} that was evaluated.
\item \code{DataSourceId} - Sets the search criteria to the \code{DataSource} used in \code{Evaluation}.
\item \code{DataUri} - Sets the search criteria to the data file(s) used in \code{Evaluation}. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.
}}

\item{EQ}{The equal to operator. The \code{Evaluation} results will have \code{FilterVariable} values that exactly match the value specified with \code{EQ}.}

\item{GT}{The greater than operator. The \code{Evaluation} results will have \code{FilterVariable} values that are greater than the value specified with \code{GT}.}

\item{LT}{The less than operator. The \code{Evaluation} results will have \code{FilterVariable} values that are less than the value specified with \code{LT}.}

\item{GE}{The greater than or equal to operator. The \code{Evaluation} results will have \code{FilterVariable} values that are greater than or equal to the value specified with \code{GE}.}

\item{LE}{The less than or equal to operator. The \code{Evaluation} results will have \code{FilterVariable} values that are less than or equal to the value specified with \code{LE}.}

\item{NE}{The not equal to operator. The \code{Evaluation} results will have \code{FilterVariable} values not equal to the value specified with \code{NE}.}

\item{Prefix}{A string that is found at the beginning of a variable, such as \code{Name} or \code{Id}.

For example, an \code{Evaluation} could have the \code{Name} \code{2014-09-09-HolidayGiftMailer}. To search for this \code{Evaluation}, select \code{Name} for the \code{FilterVariable} and any of the following strings for the \code{Prefix}:
\itemize{
\item 2014-09
\item 2014-09-09
\item 2014-09-09-Holiday
}}

\item{SortOrder}{A two-value parameter that determines the sequence of the resulting list of \code{Evaluation}.
\itemize{
\item \code{asc} - Arranges the list in ascending order (A-Z, 0-9).
\item \code{dsc} - Arranges the list in descending order (Z-A, 9-0).
}

Results are sorted by \code{FilterVariable}.}

\item{NextToken}{The ID of the page in the paginated results.}

\item{Limit}{The maximum number of \code{Evaluation} to include in the result.}
}
\description{
Returns a list of \code{DescribeEvaluations} that match the search criteria in the request.
}
\section{Accepted Parameters}{
\preformatted{describe_evaluations(
  FilterVariable = "CreatedAt"|"LastUpdatedAt"|"Status"|"Name"|"IAMUser"|"MLModelId"|"DataSourceId"|"DataURI",
  EQ = "string",
  GT = "string",
  LT = "string",
  GE = "string",
  LE = "string",
  NE = "string",
  Prefix = "string",
  SortOrder = "asc"|"dsc",
  NextToken = "string",
  Limit = 123
)
}
}

