% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{describe_tags}
\alias{describe_tags}
\title{Describes one or more of the tags for your Amazon ML object}
\usage{
describe_tags(ResourceId, ResourceType)
}
\arguments{
\item{ResourceId}{The ID of the ML object. For example, \code{exampleModelId}.}

\item{ResourceType}{The type of the ML object.}
}
\description{
Describes one or more of the tags for your Amazon ML object.
}
\examples{

}
