% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{get_ml_model}
\alias{get_ml_model}
\title{Returns an MLModel that includes detailed metadata, data source information, and the current status of the MLModel}
\usage{
get_ml_model(MLModelId, Verbose = NULL)
}
\arguments{
\item{MLModelId}{[required] The ID assigned to the \code{MLModel} at creation.}

\item{Verbose}{Specifies whether the \code{GetMLModel} operation should return \code{Recipe}.

If true, \code{Recipe} is returned.

If false, \code{Recipe} is not returned.}
}
\description{
Returns an \code{MLModel} that includes detailed metadata, data source information, and the current status of the \code{MLModel}.
}
\details{
\code{GetMLModel} provides results in normal or verbose format.
}
\section{Accepted Parameters}{
\preformatted{get_ml_model(
  MLModelId = "string",
  Verbose = TRUE|FALSE
)
}
}

