% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_package.R
\docType{package}
\name{paws.machinelearning-package}
\alias{paws.machinelearning}
\alias{paws.machinelearning-package}
\title{paws.machinelearning: Amazon Machine Learning}
\description{
Definition of the public APIs exposed by Amazon Machine
Learning
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
