% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{update_batch_prediction}
\alias{update_batch_prediction}
\title{Updates the BatchPredictionName of a BatchPrediction}
\usage{
update_batch_prediction(BatchPredictionId, BatchPredictionName)
}
\arguments{
\item{BatchPredictionId}{[required] The ID assigned to the \code{BatchPrediction} during creation.}

\item{BatchPredictionName}{[required] A new user-supplied name or description of the \code{BatchPrediction}.}
}
\description{
Updates the \code{BatchPredictionName} of a \code{BatchPrediction}.
}
\details{
You can use the \code{GetBatchPrediction} operation to view the contents of the updated data element.
}
\section{Accepted Parameters}{
\preformatted{update_batch_prediction(
  BatchPredictionId = "string",
  BatchPredictionName = "string"
)
}
}

