% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{update_data_source}
\alias{update_data_source}
\title{Updates the \code{DataSourceName} of a \code{DataSource}}
\usage{
update_data_source(
  DataSourceId = "string",
  DataSourceName = "string"
)
}
\arguments{
\item{DataSourceId}{[required] The ID assigned to the \code{DataSource} during creation.}

\item{DataSourceName}{[required] A new user-supplied name or description of the \code{DataSource} that will replace the current description.}
}
\description{
Updates the \code{DataSourceName} of a \code{DataSource}.
}
\details{
You can use the \code{GetDataSource} operation to view the contents of the updated data element.
}
\examples{

}
