% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{delete_realtime_endpoint}
\alias{delete_realtime_endpoint}
\title{Deletes a real time endpoint of an \code{MLModel}}
\usage{
delete_realtime_endpoint(MLModelId)
}
\arguments{
\item{MLModelId}{The ID assigned to the \code{MLModel} during creation.}
}
\description{
Deletes a real time endpoint of an \code{MLModel}.
}
\examples{

}
