% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{get_batch_prediction}
\alias{get_batch_prediction}
\title{Returns a \code{BatchPrediction} that includes detailed metadata, status, and data file information for a \code{Batch Prediction} request}
\usage{
get_batch_prediction(BatchPredictionId)
}
\arguments{
\item{BatchPredictionId}{An ID assigned to the \code{BatchPrediction} at creation.}
}
\description{
Returns a \code{BatchPrediction} that includes detailed metadata, status, and data file information for a \code{Batch Prediction} request.
}
\examples{

}
