% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.machinelearning_operations.R
\name{predict}
\alias{predict}
\title{Generates a prediction for the observation using the specified \code{ML Model}}
\usage{
predict(
  MLModelId = "string",
  Record = list(
    "string"
  ),
  PredictEndpoint = "string"
)
}
\arguments{
\item{MLModelId}{[required] A unique identifier of the \code{MLModel}.}

\item{Record}{[required]}

\item{PredictEndpoint}{[required]}
}
\description{
Generates a prediction for the observation using the specified \code{ML Model}.
}
\details{
Note

Not all response parameters will be populated. Whether a response parameter is populated depends on the type of model requested.
}
\examples{

}
